+++ package.json
@@ -18,12 +18,16 @@
   },
   "dependencies": {
     "@react-native-community/async-storage": "^1.12.1",
+    "@react-navigation/native": "^6.0.8",
+    "@react-navigation/native-stack": "^6.4.1",
     "i18next": "^21.6.10",
     "prop-types": "^15.8.1",
     "react": "17.0.2",
     "react-i18next": "^11.15.3",
     "react-native": "0.67.2",
-    "react-native-localize": "^2.1.9"
+    "react-native-localize": "^2.1.9",
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-screens": "^3.10.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
