+++ src/navigation/AppStack.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import HomeStack from './stack/HomeStack';
+
+const AppStack = () => {
+  return <HomeStack />;
+};
+
+export default AppStack;
