+++ src/navigation/AuthStack.js
@@ -1,0 +1,12 @@
+import {View, Text} from 'react-native';
+import React from 'react';
+
+const AuthStack = () => {
+  return (
+    <View>
+      <Text>Auth Stack</Text>
+    </View>
+  );
+};
+
+export default AuthStack;
