+++ src/navigation/stack/HomeStack.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import HomeScreen from '@screens/Home';
+
+const HomeStack = () => {
+  return <HomeScreen />;
+};
+
+export default HomeStack;
