+++ src/screens/Home/index.js
@@ -1,0 +1,3 @@
+import HomeScreen from './Home.Screen';
+
+export default HomeScreen;
